app/requirements.txt
attrs==25.1.0
Automat==0.8.0
beautifulsoup4==4.13.3
blinker==1.8.2
cached-property==1.5.1
certifi==2025.1.31
chardet==3.0.4
click==8.1.8
cloud-init==24.1.3
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
distro==1.4.0
distro-info==0.23+ubuntu1.1
docker==4.1.0
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
exceptiongroup==1.2.2
flask==3.0.3
greenlet==3.0.3
h11==0.14.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==8.5.0
incremental==16.10.1
itsdangerous==2.2.0
jinja2==3.1.5
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lxml==5.3.1
MarkupSafe==2.1.5
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
outcome==1.3.0.post0
pexpect==4.6.0
playwright==1.42.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pyee==11.0.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
PySocks==1.7.1
python-apt==2.0.1+ubuntu0.20.4.1
python-debian==0.1.36+ubuntu1.1
python-dotenv==1.0.1
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sniffio==1.3.1
sortedcontainers==2.4.0
sos==4.5.6
soupsieve==2.6
ssh-import-id==5.10
systemd-python==234
texttable==1.6.2
trio==0.27.0
trio-websocket==0.12.1
Twisted==18.9.0
typing-extensions==4.12.2
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==2.2.3
wadllib==1.3.3
websocket-client==1.8.0
werkzeug==3.0.6
wsproto==1.2.0
zipp==3.20.2
zope.interface==4.7.1
